Move LibFiles tests to config tests
